[CCUBE-1510-2][JY] Refactor basic details and pills test cases
