+++ src/components/element-editor/basic-details/dropdown/preselected-value-dropdown.tsx
@@ -15,8 +15,6 @@ export const PreselectedValueDropdown = () => {
         formState: { errors },
         watch,
         setValue,
-        getValues,
-        resetField,
     } = useFormContext<TOverallOptionGroupBasedSchema>();
 
     const preselectedValue = watch("preselectedValue");
