+++ src/components/element-editor/basic-details/email/email-basic-details.tsx
@@ -9,7 +9,6 @@ import {
 } from "../common";
 
 export const EmailBasicDetails = () => {
-    console.log("EmailBasicDetails");
     return (
         <>
             <Type />
